# [RFC] Use 6to5 and take advantage of rest transforms (avoids leaking arguments)