The is mostly a request for comment as I understand moving to 6to5 might have other implications. But tests all pass and performance improves in some cases. Thoughts?

The rest params transform looks a lot like your `arrCopy` (minus the leaking of args because it is inline):

```
for (var _len4 = arguments.length, values = Array(_len4), _key4 = 0; _key4 < _len4; _key4++) {
    values[_key4] = arguments[_key4];
}
```

```
Running "stats" task
  Raw: 142,546 bytes         +4,050 bytes
  Min:  58,247 bytes 59.13%  +2,215 bytes
  Zip:  15,823 bytes 88.89%    +122 bytes
```

```
List > builds from array of 2
  Old:   691,602   699,059   706,679 ops/sec
  New:   808,152   815,852   823,701 ops/sec
  compare: 1 1
  diff: 16.7%
  rme: 1.01%
List > builds from array of 8
  Old:   640,498   645,867   651,326 ops/sec
  New:   740,372   746,932   753,609 ops/sec
  compare: 1 1
  diff: 15.64%
  rme: 0.86%
List > builds from array of 32
  Old:   254,254   256,676   259,145 ops/sec
  New:   272,485   274,272   276,083 ops/sec
  compare: 1 1
  diff: 6.85%
  rme: 0.81%
List > builds from array of 1024
  Old:     9,785     9,870     9,957 ops/sec
  New:     9,848     9,930    10,012 ops/sec
  compare: 1 1
  diff: 0.6%
  rme: 0.85%
List > pushes into 2 times
  Old:   839,622   844,886   850,217 ops/sec
  New:   877,473   885,676   894,033 ops/sec
  compare: 1 1
  diff: 4.82%
  rme: 0.79%
List > pushes into 8 times
  Old:   189,652   191,290   192,956 ops/sec
  New:   191,637   196,253   201,097 ops/sec
  compare: 1 1
  diff: 2.59%
  rme: 1.8%
List > pushes into 32 times
  Old:    45,151    45,543    45,941 ops/sec
  New:    45,781    46,111    46,446 ops/sec
  compare: 1 1
  diff: 1.24%
  rme: 0.79%
List > pushes into 1024 times
  Old:     1,394     1,406     1,419 ops/sec
  New:     1,475     1,487     1,499 ops/sec
  compare: 1 1
  diff: 5.7%
  rme: 0.84%
List > pushes into transient 2 times
  Old: 1,328,763 1,341,117 1,353,704 ops/sec
  New: 1,478,902 1,489,587 1,500,428 ops/sec
  compare: 1 1
  diff: 11.07%
  rme: 0.83%
List > pushes into transient 8 times
  Old:   525,532   530,718   536,007 ops/sec
  New:   534,015   538,823   543,719 ops/sec
  compare: 1 1
  diff: 1.52%
  rme: 0.94%
List > pushes into transient 32 times
  Old:   162,114   163,528   164,967 ops/sec
  New:   162,803   164,121   165,460 ops/sec
  compare: 0 0
  diff: 0.36%
  rme: 0.84%
List > pushes into transient 1024 times
  Old:     5,002     5,130     5,265 ops/sec
  New:     5,281     5,329     5,379 ops/sec
  compare: 1 1
  diff: 3.88%
  rme: 1.92%
Map builds from an object of 2
  Old:   531,265   536,883   542,622 ops/sec
  New:   480,687   485,541   490,495 ops/sec
  compare: 1 1
  diff: -9.57%
  rme: 1.03%
Map builds from an object of 8
  Old:   105,747   106,688   107,646 ops/sec
  New:   101,170   102,090   103,026 ops/sec
  compare: 1 1
  diff: -4.32%
  rme: 0.89%
Map builds from an object of 32
  Old:    37,285    37,653    38,028 ops/sec
  New:    37,252    37,552    37,856 ops/sec
  compare: 0 0
  diff: -0.27%
  rme: 0.9%
Map builds from an object of 1024
  Old:     1,836     1,852     1,868 ops/sec
  New:     1,758     1,801     1,845 ops/sec
  compare: 1 1
  diff: -2.76%
  rme: 1.81%
all done
```