@leebyron 2015-01-26
Yes! I actually have an almost identical change sitting in a branch on my local machine to do this, and I've been building out the perf tests in order to measure exactly the question you're asking about implications.

> The rest params transform looks a lot like your arrCopy (minus the leaking of args because it is inline):

Yep! I added that change to 6to5!

So I'm all for 6to5, but just need more time to understand the implications first.