@sebmck 2015-02-01
Would be interested to find out the cases that 6to5 has actually made slower. Looks like definite areas for improvement.