@camspiers 2015-02-02
@sebmck just speculating, but I think that it is the conversion from the array-like `arguments` to array. This is good as it avoids leaking args, but it isn't without cost.

https://github.com/petkaantonov/bluebird/wiki/Optimization-killers#3-managing-arguments