@sebmck 2015-02-02
@camspiers As @leebyron said engines optimise for `fn.apply(this, arguments);` and doesn't leak.