@camspiers 2015-02-02
A lot of the changes don't use `apply`, but yes you are definitely right that some of the changes are unnecessary (and should be changed). Thanks heaps for the information :+1: