@sebmck 2015-02-02
@camspiers Ah right sorry, I just skimmed over the changes. Looks good though! Let me know if you find any transformation output that can be optimised. Might be worth taking a look at [loose mode](http://6to5.org/docs/usage/loose/) if immutable uses any of the features specified on that page.